pyproject.toml
[project]
name = "spirits"
version = "1.1.0"
authors = [{ name = "Hugh Liu", email = "hugh.lpd@gmail.com" }]
description = "spirits maker and gather via CLI or script"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.6"
classifiers = ["Typing :: Typed", "Programming Language :: Python :: 3 :: Only"]
dependencies = ["requests", "pillow", "typer"]

[project.urls]
Repository = "https://github.com/HUGHNew/spirits"

[project.scripts]
spirits = "spirits.command:spirits_cmd"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
